Change arguments to imx5x-mmc-prep.sh
* Remove default SD-card device. Defaulting to something like /dev/sdf
  looks a bit too dangerous to me...
* Figure out the rootfs partition device from the SD-card device
* Remove default image arguments, and support updating only some of the
  images on the card.

Also, set x bit on the script. (closes #381)

+++ scripts/imx5x-mmc-prep.sh
@@ -1,18 +1,15 @@
 #!/bin/bash 
-DEVICE="${1-${DEVICE-/dev/sdf}}"
-[ -a "${DEVICE}" ] || exit 1
-shift
-PART_ROOTFS="${1-${PART_ROOTFS-/dev/sdf1}}"
-shift
-UBOOT="${1-${UBOOT-u-boot.bin}}"
-[ -r "${UBOOT}" ] || exit 1
-shift
-KERNEL="${1-${KERNEL-uImage}}"
-shift
-ROOTFS="${1-${ROOTFS-rootfs.tar.gz}}"
-shift
 
-set -e -x
+# Call this script with something like:
+# imx5x-mmc-prep.sh /dev/mmcblk0 \
+#	tmp/images/imx53qsb-u-boot-imx.bin \
+#	tmp/images/imx53qsb-linux-imx.bin \
+#	tmp/images/imx53qsb-base-rootfs.tar.gz
+
+DEVICE=$1
+UBOOT=$2
+KERNEL=$3
+ROOTFS=$4
 
 dev_umount() {
     for i in $( grep ${1} /proc/mounts | awk '{print $1}' );
@@ -21,33 +18,52 @@ dev_umount() {
     done
 }
 
-#######################
-ls -l $DEVICE
-sleep 5
+set -e -x
+
+# sd-card device argument must be specified
+[ -n "$DEVICE" ]
+# and device must be writable
+test -w $DEVICE
 
 dev_umount $DEVICE
 
+if [ -n "$UBOOT" ] ; then
 dd if=$UBOOT of=$DEVICE bs=1k seek=1 skip=1
+fi
 
-[ -r $KERNEL ]
+if [ -n "$KERNEL" ] ; then
 dd if=$KERNEL of=$DEVICE bs=1k seek=1024 skip=0
+fi
 
-[ -r $ROOTFS ]
+if [ -n "$ROOTFS" ] ; then
+
 dd if=/dev/zero of=$DEVICE bs=512 count=1
 partprobe $DEVICE
 
 sfdisk -uM $DEVICE <<EOF
 8,16,83
 EOF
-
 sleep 3
-
 dev_umount $DEVICE
 
-mkfs.ext4 -L rootfs $PART_ROOTFS
-mkdir -p /mnt/tmp-sdrootfs
-mount -t ext4 $PART_ROOTFS /mnt/tmp-sdrootfs
-tar -xf $ROOTFS -C /mnt/tmp-sdrootfs
+# find rootfs partition
+if [ -b "${DEVICE}p1" ]; then
+	ROOTFS_DEVICE=${DEVICE}p1
+elif [ -b "${DEVICE}1" ]; then
+	ROOTFS_DEVICE=${DEVICE}1
+else
+	echo "Could not find device"
+	exit 1;
+fi
+
+mkfs.ext4 -L rootfs $ROOTFS_DEVICE
+ROOTFS_MNT=`mktemp -d`
+mount -t ext4 $ROOTFS_DEVICE $ROOTFS_MNT
+tar -xf $ROOTFS -C $ROOTFS_MNT
 sleep 5
-umount /mnt/tmp-sdrootfs
-rm -rf /mnt/tmp-sdrootfs
+umount $ROOTFS_MNT
+rm -rf $ROOTFS_MNT
+
+fi
+
+sync
